UMAP - add color-by data to anndata.obs in new step, then pull from here for plotting
