+++ vulcan/lib/server/workflows/scripts/umap_prep_plotting_data.py
@@ -1,29 +1,14 @@
 from archimedes.functions.dataflow import output_path, output_json, input_path, input_var, input_json, buildTargetPath
 from archimedes.functions.scanpy import scanpy as sc
-from archimedes.functions.magma import connect, question
 from archimedes.functions.list import flatten
 from archimedes.functions.environment import project_name
 from archimedes.functions.utils import pandas as pd
+from archimedes.functions.utils import re
 
 # Read inputs
-scdata = sc.read(input_path('umap_anndata.h5ad'))
-
-leiden = list(map(str, input_json('leiden.json')))
+scdata = sc.read(input_path('scdata.h5ad'))
 
 color_by = input_var('color_by')
-
-# Prep magma querying
-magma = connect()
-def get(ids, value):
-    ids = ids if type(ids) == list else ids.tolist()
-    values = dict(question(magma, [
-        'sc_seq',
-        [ '::identifier', '::in', ids ],
-        '::all',
-        *value
-    ], strip_identifiers=False))
-
-    return [ values.get(id, None) for id in ids ]
 
 pdat = input_json("project_data")[project_name]
 color_options = pdat['color_options']
@@ -31,7 +16,7 @@ color_options = pdat['color_options']
 # Obtain color data
 custom_tooltip = False
 if color_by == 'Cluster':
-    color = leiden
+    color = scdata.obs[ 'leiden' ].values
     custom_tooltip = True
     sets = input_json('top10.json')
     texts = dict([
@@ -40,16 +25,15 @@ if color_by == 'Cluster':
             ' ' + ', '.join(sets[clust])
         ] for clust in list(sets.keys()) ])
     hover_name = 'top10 markers'
-    hover_text = [texts[str(val)] for val in leiden]
+    hover_text = [texts[str(val)] for val in color]
 elif color_by == 'Manual Annotations':
-    annots = input_json('annots.json')
-    color = [annots[x] for x in leiden]
+    color = scdata.obs[ 'Manual_Annotations' ].values
 elif color_by == 'Tube':
     color = scdata.obs[ 'Record_ID' ].values
+elif color_by in color_options.keys():
+    color = scdata.obs[ re.sub(" ", "_", color_by) ].values
 elif color_by in scdata.raw.var_names:
     color = flatten(scdata.raw.X[ : , scdata.raw.var_names == color_by ].toarray())
-elif color_by in color_options.keys():
-    color = get(scdata.obs[ 'Record_ID' ].values, buildTargetPath(color_options[color_by], pdat))
 else:
     color = None
